requirements.txt
keyring
qrcode
pyserial